+++ packages/app/src/specs/tree/tree.spec.ts
@@ -1,0 +1,8 @@
+import { it } from 'vitest'
+import { } from './useCollapsibleTree'
+
+describe('...', () => {
+  it('works', () => {
+
+  })
+})
